+++ src/infrastructure/openai/openai.service.ts
