+++ src/infrastructure/openai/prompt.ts
